+++ lib/release.js
@@ -6,6 +6,8 @@ module.exports = function zipCurrentDirectory(cb) {
 	try {
 		fs.unlinkSync('./archive.interactive')
 	} catch(err) {}
-	zipFolder(process.cwd(), './archive.interactive', cb)
-	console.log(chalk.green('Created archive.interactive.'))
+	zipFolder(process.cwd(), './archive.interactive', function() {
+		console.log(chalk.green('Created archive.interactive.'))
+		cb()
+	})
 }
