+++ lib/server.js
@@ -1,7 +1,6 @@
 var browserSync = require('browser-sync')
 var fs = require('fs')
 var path = require('path')
-// var guid = require('./guid')()
 var logClient = fs.readFileSync(path.join(__dirname, 'public/logClient.js'), { encoding: 'UTF-8' })
 
 module.exports = function(options) {
@@ -20,7 +19,7 @@ module.exports = function(options) {
 	browserSync({
 		files: './**',
 		https: true,
-		open: options.open,
+		open: 'external',
 		startPath: '/' + mcode,
 		rewriteRules: [{
 			match: /<head>/i,
@@ -39,7 +38,4 @@ module.exports = function(options) {
 	}, function(err, bs) {
 		require('./socketLogger')(bs.io.sockets)
 	})
-
-	// console.log('For weinre support, add this script tag in the index.html of your Interactive', '<script src="https://weinre.mybluemix.net/target/target-script-min.js#' + guid + '"></script>')
-	// console.log('Open URL in your browser: https://weinre.mybluemix.net/client#' + guid)
 }
