+++ lib/uploader.js
@@ -5,19 +5,6 @@ var chalk = require('chalk')
 
 var fileSizeInBytes = 0;
 
-function getAccessToken(username, password, cb) {
-	var url = 'https://accounts.mediafly.com/api/3.3/authentication/authenticate?accessType=edit&username=' + 
-			encodeURIComponent(username) + '&password= ' + encodeURIComponent(password)
-
-	request.get({url: url, json: true}, function(err, response, body) {
-			if (err) {
-				console.log(chalk.red('Error authenticating. Please check your credentials.'))
-				return
-			}
-			cb(body.accessToken)
-		})
-}
-
 function getPresignedUrl(accessToken, productId, cb) {
 	request.get({url: 'https://launchpadapi.mediafly.com/uploads/signedurl?accessToken=' + accessToken +
 			'&productid=' + productId, json: true}, function(err, response, body) {
@@ -64,26 +51,22 @@ function updateAsset(productId, accessToken, itemId, s3Url, cb) {
 	}, cb)
 }
 
-module.exports = function(userId, password, productId, itemId) {
+module.exports = function(accessToken, productId, itemId) {
 	//zip up the current directory
-	var accessToken
 	release(function() {
-		getAccessToken(userId, password, function(token) {
-			accessToken = token
-			getPresignedUrl(accessToken, productId, function(urlResponse) {
-				uploadInteractive(urlResponse.signedUrl, function(err) {
+		getPresignedUrl(accessToken, productId, function(urlResponse) {
+			uploadInteractive(urlResponse.signedUrl, function(err) {
+				if (err) {
+					console.log(chalk.red('Error uploading interactive.'))
+					return
+				}
+
+				updateAsset(productId, accessToken, itemId, urlResponse.url, function(err) {
 					if (err) {
-						console.log(chalk.red('Error uploading interactive.'))
+						console.log(chalk.red('Error updating '))
 						return
 					}
-
-					updateAsset(productId, accessToken, itemId, urlResponse.url, function(err) {
-						if (err) {
-							console.log(chalk.red('Error updating '))
-							return
-						}
-						console.log(chalk.green('Interactive successfully updated.'))
-					})
+					console.log(chalk.green('Interactive successfully updated.'))
 				})
 			})
 		})
